# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AccountBook API 설계 명세",
    "version": "1.2.0",
    "description": "인증·가계부·소비 분석·OCR에서 구현해야 할 API 계약의 단일 기준인 OpenAPI 3.1 문서입니다. 2026-09-18 현재 Python 서버는 공통 code 응답과 계정 PATCH·DELETE를 포함한 인증 API를 구현했습니다. 월별 예산 분리, 거래·리포트·OCR API는 아직 구현되지 않았습니다. 구현이 완료될 때까지 실행 중 서버의 /docs와 /openapi.json은 현재 구현 상태를 보여주고, 이 파일은 목표 계약을 나타냅니다.\n\n날짜·일시 형식 규칙: 달력 날짜는 ISO 8601 YYYY-MM-DD, 정확한 시점은 한국 시간 오프셋을 포함한 YYYY-MM-DDTHH:mm:ss+09:00을 사용합니다. 거래 시각처럼 날짜와 분리된 선택 시각은 HH:mm:ss 또는 null을 사용합니다."
  },
  "paths": {
    "/api/v1/auth/sign-up": {
      "post": {
        "tags": [
          "사용자 관리 및 인증"
        ],
        "summary": "사용자 등록",
        "operationId": "post_api_v1_auth_sign_up",
        "description": "필수 JSON: `username`, `password`, `display_name`.\n\n```json\n{\"username\":\"User123\",\"password\":\"Password1!\",\"display_name\":\"홍길동\"}\n```\n\n성공 200의 data는 null입니다. 입력이 잘못되면 오류 코드 400, 아이디가 중복되면 오류 코드 409, DB 오류가 발생하면 오류 코드 503을 반환합니다. 가입 성공으로 로그인 토큰을 발급하지 않습니다.\n\n공통 계약\n\nPOST·PUT·PATCH는 `Content-Type: application/json`을 사용합니다. 알 수 없는 필드, 잘못된 자료형 또는 허용되지 않은 쿼리 파라미터를 전달하거나 GET 요청에 본문을 포함하면 오류 코드 400을 반환합니다. 오류 응답에 입력값·비밀번호·토큰·DB 경로를 포함하지 않습니다. 응답에는 `Cache-Control: no-store`를 설정합니다.\n\n모든 API 응답은 HTTP 상태와 같은 `status`, 안정적인 영문 `code`, 영문 `message`, `data`를 포함합니다. 오류의 `data`는 null이며 클라이언트는 message가 아니라 code로 분기합니다.\n\n```json\n{\"status\":200,\"code\":\"SUCCESS\",\"message\":\"The request has been accepted and processed.\",\"data\":null}\n```\n\n| 상태 | code | message | 의미 |\n| --- | --- | --- | --- |\n| 200 | `SUCCESS` | The request has been accepted and processed. | 성공 |\n| 400 | `INVALID_REQUEST` | The request parameters or format are invalid. | 요청 형식·입력 정책 위반 |\n| 401 | `INVALID_CREDENTIALS` | The username or password is incorrect. | 로그인 실패; 아이디 존재 여부를 구분하지 않음 |\n| 401 | `AUTHENTICATION_REQUIRED` | Authentication is required or the token is invalid. | 토큰 누락·검증 실패·만료·폐기 |\n| 401 | `CURRENT_PASSWORD_INCORRECT` | The current password is incorrect. | 비밀번호 변경 시 현재 비밀번호 불일치 |\n| 409 | `USERNAME_CONFLICT` | The username is already in use. | 가입 아이디 중복 |\n| 429 | `SIGN_IN_RATE_LIMITED` | Too many sign-in attempts. Please try again later. | 로그인 일시 제한; Retry-After는 남은 초 |\n| 503 | `SERVICE_UNAVAILABLE` | The service is temporarily unavailable. | DB 잠금·일시 사용 불가 |\n\n401 응답에는 `WWW-Authenticate: Bearer`를 설정합니다. 위 503은 DB 오류 응답이며 Uvicorn의 동시 요청 제한 응답까지 이 JSON 형식을 보장하지는 않습니다.\n\n토큰 수명과 보안\n\nJWT는 HS256으로 서명하며 기본 유효기간은 발급부터 5일입니다. sub(내부 사용자 ID), ver(폐기 버전), jti(고유 ID), iat, exp, iss, aud를 검증합니다. 사용자가 존재하고 활성 상태이며 DB의 버전과 일치해야 합니다. 비밀번호는 Argon2id 해시로 저장합니다.\n\n갱신은 아직 유효한 JWT로 새 발급 시각·고유 ID·유효기간의 JWT를 생성합니다. 갱신 전 토큰도 원래 만료 시각까지 유효합니다. 별도 리프레시 토큰, 30일 누적 갱신 상한, 로그인 세션 테이블은 없습니다. 프론트엔드 권장 조건은 **발급 후 1일 이상 경과 AND 만료까지 6시간 이내**입니다. 서버는 이 권장 시점을 강제하지 않습니다. 만료 후에는 다시 로그인합니다.\n\n클라이언트는 JWT를 안전하게 보관하고 요청 위치 규칙을 따릅니다. 로그아웃 API는 없으며 클라이언트에서 토큰을 삭제해도 서버에서 그 토큰을 폐기하지 않습니다. 실제 배포에서는 HTTPS로 평문 비밀번호와 토큰을 보호해야 합니다.\n\n조회·수정은 인증된 본인 계정에만 적용합니다.",
        "responses": {
          "200": {
            "description": "정상 처리됨.",
            "headers": {
              "Cache-Control": {
                "description": "캐시 저장 금지",
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 200
                    },
                    "message": {
                      "type": "string",
                      "const": "The request has been accepted and processed."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 200,
                  "message": "The request has been accepted and processed.",
                  "data": null,
                  "code": "SUCCESS"
                }
              }
            }
          },
          "400": {
            "description": "The request parameters or format are invalid.",
            "headers": {
              "Cache-Control": {
                "description": "캐시 저장 금지",
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 400
                    },
                    "message": {
                      "type": "string",
                      "const": "The request parameters or format are invalid."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 400,
                  "message": "The request parameters or format are invalid.",
                  "data": null,
                  "code": "INVALID_REQUEST"
                }
              }
            }
          },
          "409": {
            "description": "The username is already in use.",
            "headers": {
              "Cache-Control": {
                "description": "캐시 저장 금지",
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 409
                    },
                    "message": {
                      "type": "string",
                      "const": "The username is already in use."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 409,
                  "message": "The username is already in use.",
                  "data": null,
                  "code": "USERNAME_CONFLICT"
                }
              }
            }
          },
          "503": {
            "description": "The service is temporarily unavailable. DB 잠금·일시 사용 불가. Uvicorn의 동시 요청 제한 응답까지 이 JSON 형식을 보장하지는 않습니다.",
            "headers": {
              "Cache-Control": {
                "description": "캐시 저장 금지",
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 503
                    },
                    "message": {
                      "type": "string",
                      "const": "The service is temporarily unavailable."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 503,
                  "message": "The service is temporarily unavailable.",
                  "data": null,
                  "code": "SERVICE_UNAVAILABLE"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "application/json 본문은 필수입니다. 쿼리 파라미터나 알 수 없는 필드를 전달하면 오류 코드 400을 반환합니다.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 4,
                    "maxLength": 12,
                    "pattern": "^[A-Za-z0-9]{4,12}(?![\\s\\S])",
                    "description": "아이디: 4–12자, ASCII 영문·숫자만 허용합니다. 소문자로 저장하여 대소문자를 구분하지 않습니다. 공백을 자동 제거하지 않습니다."
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 20,
                    "format": "password",
                    "writeOnly": true,
                    "pattern": "^(?:(?=.*[A-Za-z])(?=.*[0-9])|(?=.*[A-Za-z])(?=.*[!@#$%^&*_=+?\\-])|(?=.*[0-9])(?=.*[!@#$%^&*_=+?\\-]))[A-Za-z0-9!@#$%^&*_=+?\\-]{8,20}(?![\\s\\S])",
                    "description": "비밀번호: 8–20자, `A–Z`, `a–z`, `0–9`, `!@#$%^&*_-+=?`만 허용합니다. 영문·숫자·특수문자 세 종류 중 두 종류 이상이 필요합니다. 대소문자를 구분하며 공백·한글·그 외 문자를 거부합니다. 가입·비밀번호 변경에 이 정책을 적용하며, 로그인은 입력한 비밀번호를 해시와 비교합니다."
                  },
                  "display_name": {
                    "type": "string",
                    "pattern": "^ *[A-Za-z0-9가-힣ㄱ-ㆎᄀ-ᇿ][A-Za-z0-9가-힣ㄱ-ㆎᄀ-ᇿ ]{0,8}[A-Za-z0-9가-힣ㄱ-ㆎᄀ-ᇿ] *(?![\\s\\S])",
                    "description": "표시 이름: 일반 공백을 앞뒤에서 제거한 후 2–10자입니다. 한글 음절·자모, ASCII 영문·숫자·일반 공백만 허용합니다. 중간 공백을 유지하고 글자 수에 포함합니다. 공백만 있는 이름·탭·줄바꿈·제어문자·다른 특수문자를 거부합니다. 중복은 허용합니다. 원본 입력은 앞뒤 공백을 포함하여 최대 256자이며, 공백 제거 후 2–10자 제한은 pattern으로 검사합니다.",
                    "maxLength": 256
                  }
                },
                "required": [
                  "username",
                  "password",
                  "display_name"
                ]
              },
              "example": {
                "username": "User123",
                "password": "Password1!",
                "display_name": "홍길동"
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "사용자 관리 및 인증"
        ],
        "summary": "사용자 토큰 갱신",
        "operationId": "post_api_v1_auth_refresh",
        "description": "필수 JSON: `{\"token\":\"<JWT>\"}`. Authorization 헤더를 함께 보내거나 헤더로 대체하면 오류 코드 400을 반환합니다.\n\n성공 200의 data는 token, expires_at, token_type을 포함합니다. token을 누락하거나 요청 형식이 잘못되면 오류 코드 400, token 검증에 실패하거나 이미 만료된 token을 전달하면 오류 코드 401, DB 오류가 발생하면 오류 코드 503을 반환합니다.\n\n공통 계약\n\nPOST·PUT·PATCH는 `Content-Type: application/json`을 사용합니다. 알 수 없는 필드, 잘못된 자료형 또는 허용되지 않은 쿼리 파라미터를 전달하거나 GET 요청에 본문을 포함하면 오류 코드 400을 반환합니다. 오류 응답에 입력값·비밀번호·토큰·DB 경로를 포함하지 않습니다. 응답에는 `Cache-Control: no-store`를 설정합니다.\n\n모든 API 응답은 HTTP 상태와 같은 `status`, 안정적인 영문 `code`, 영문 `message`, `data`를 포함합니다. 오류의 `data`는 null이며 클라이언트는 message가 아니라 code로 분기합니다.\n\n```json\n{\"status\":200,\"code\":\"SUCCESS\",\"message\":\"The request has been accepted and processed.\",\"data\":null}\n```\n\n| 상태 | code | message | 의미 |\n| --- | --- | --- | --- |\n| 200 | `SUCCESS` | The request has been accepted and processed. | 성공 |\n| 400 | `INVALID_REQUEST` | The request parameters or format are invalid. | 요청 형식·입력 정책 위반 |\n| 401 | `INVALID_CREDENTIALS` | The username or password is incorrect. | 로그인 실패; 아이디 존재 여부를 구분하지 않음 |\n| 401 | `AUTHENTICATION_REQUIRED` | Authentication is required or the token is invalid. | 토큰 누락·검증 실패·만료·폐기 |\n| 401 | `CURRENT_PASSWORD_INCORRECT` | The current password is incorrect. | 비밀번호 변경 시 현재 비밀번호 불일치 |\n| 409 | `USERNAME_CONFLICT` | The username is already in use. | 가입 아이디 중복 |\n| 429 | `SIGN_IN_RATE_LIMITED` | Too many sign-in attempts. Please try again later. | 로그인 일시 제한; Retry-After는 남은 초 |\n| 503 | `SERVICE_UNAVAILABLE` | The service is temporarily unavailable. | DB 잠금·일시 사용 불가 |\n\n401 응답에는 `WWW-Authenticate: Bearer`를 설정합니다. 위 503은 DB 오류 응답이며 Uvicorn의 동시 요청 제한 응답까지 이 JSON 형식을 보장하지는 않습니다.\n\n토큰 수명과 보안\n\nJWT는 HS256으로 서명하며 기본 유효기간은 발급부터 5일입니다. sub(내부 사용자 ID), ver(폐기 버전), jti(고유 ID), iat, exp, iss, aud를 검증합니다. 사용자가 존재하고 활성 상태이며 DB의 버전과 일치해야 합니다. 비밀번호는 Argon2id 해시로 저장합니다.\n\n갱신은 아직 유효한 JWT로 새 발급 시각·고유 ID·유효기간의 JWT를 생성합니다. 갱신 전 토큰도 원래 만료 시각까지 유효합니다. 별도 리프레시 토큰, 30일 누적 갱신 상한, 로그인 세션 테이블은 없습니다. 프론트엔드 권장 조건은 **발급 후 1일 이상 경과 AND 만료까지 6시간 이내**입니다. 서버는 이 권장 시점을 강제하지 않습니다. 만료 후에는 다시 로그인합니다.\n\n클라이언트는 JWT를 안전하게 보관하고 요청 위치 규칙을 따릅니다. 로그아웃 API는 없으며 클라이언트에서 토큰을 삭제해도 서버에서 그 토큰을 폐기하지 않습니다. 실제 배포에서는 HTTPS로 평문 비밀번호와 토큰을 보호해야 합니다.\n\n조회·수정은 인증된 본인 계정에만 적용합니다.\n기본 JWT 유효기간은 TOKEN_SECONDS=432000이며 서버 설정으로 변경할 수 있습니다.",
        "responses": {
          "200": {
            "description": "정상 처리됨.",
            "headers": {
              "Cache-Control": {
                "description": "캐시 저장 금지",
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 200
                    },
                    "message": {
                      "type": "string",
                      "const": "The request has been accepted and processed."
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "token",
                        "expires_at",
                        "token_type"
                      ],
                      "properties": {
                        "token": {
                          "type": "string",
                          "description": "새로 발급한 JWT입니다."
                        },
                        "expires_at": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T(?:[01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]\\+09:00(?![\\s\\S])",
                          "example": "2026-09-16T14:30:00+09:00",
                          "description": "발급한 JWT가 만료되는 한국 시간입니다. JWT의 exp와 같은 시점입니다."
                        },
                        "token_type": {
                          "type": "string",
                          "const": "Bearer",
                          "description": "Authorization 헤더에 사용할 인증 유형입니다."
                        }
                      }
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 200,
                  "message": "The request has been accepted and processed.",
                  "data": {
                    "token": "<새 JWT>",
                    "expires_at": "2026-09-21T14:30:00+09:00",
                    "token_type": "Bearer"
                  },
                  "code": "SUCCESS"
                }
              }
            }
          },
          "400": {
            "description": "The request parameters or format are invalid.",
            "headers": {
              "Cache-Control": {
                "description": "캐시 저장 금지",
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 400
                    },
                    "message": {
                      "type": "string",
                      "const": "The request parameters or format are invalid."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 400,
                  "message": "The request parameters or format are invalid.",
                  "data": null,
                  "code": "INVALID_REQUEST"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or the token is invalid.",
            "headers": {
              "Cache-Control": {
                "description": "캐시 저장 금지",
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "WWW-Authenticate": {
                "description": "Bearer 인증 안내",
                "schema": {
                  "type": "string",
                  "const": "Bearer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 401
                    },
                    "message": {
                      "type": "string",
                      "const": "Authentication is required or the token is invalid."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 401,
                  "message": "Authentication is required or the token is invalid.",
                  "data": null,
                  "code": "AUTHENTICATION_REQUIRED"
                }
              }
            }
          },
          "503": {
            "description": "The service is temporarily unavailable. DB 잠금·일시 사용 불가. Uvicorn의 동시 요청 제한 응답까지 이 JSON 형식을 보장하지는 않습니다.",
            "headers": {
              "Cache-Control": {
                "description": "캐시 저장 금지",
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 503
                    },
                    "message": {
                      "type": "string",
                      "const": "The service is temporarily unavailable."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 503,
                  "message": "The service is temporarily unavailable.",
                  "data": null,
                  "code": "SERVICE_UNAVAILABLE"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "application/json 본문은 필수입니다. 쿼리 파라미터나 알 수 없는 필드를 전달하면 오류 코드 400을 반환합니다.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 1,
                    "description": "요청 JSON에 첨부할 유효한 JWT입니다. 토큰 수명·검증 정책은 API 설명을 참고합니다.",
                    "writeOnly": true,
                    "maxLength": 4096
                  }
                },
                "required": [
                  "token"
                ]
              },
              "example": {
                "token": "<JWT>"
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authorization: Bearer <JWT> 헤더로 인증합니다."
      }
    },
...
  }
}